requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.23
pydantic>=2.5.2
redis>=5.0.1
cassandra-driver>=3.28.0
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation>=0.42b0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-proto>=1.21.0
python-dotenv>=1.0.0
langchain>=0.0.350
requests>=2.31.0
backoff>=2.2.1
bcrypt>=4.1.2

# Testing dependencies
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.6.1
httpx>=0.26.0
aioresponses>=0.7.6
freezegun>=1.4.0
